+++ toolchain/scripts/modules/assert/rule.ps1
@@ -25,6 +25,7 @@ function Assert-Rule {
     if ($Rule -eq "serve:mame") { return 0 }
     if ($Rule -like "run:mame*") { return 0 }
     if ($Rule -eq "run:raine") {return 0 }
+    if ($Rule -like "run:raine:*") { return 0 }
 
     if ($Rule -eq "run") { # TODO : remove at v3
       Write-Host "warning : run rule is deprecated, use run:mame instead" -ForegroundColor Yellow
@@ -75,8 +76,6 @@ function Assert-Rule {
     Write-Host "run"
     Write-Host "run:mame"
     Write-Host "run:raine"
-    Write-Host "--version"
-    Write-Host "--help"
     Write-Host ""
 
     if ($Config.project.emulator.mame.profile) {
@@ -86,6 +85,16 @@ function Assert-Rule {
         Write-Host "run:mame:$($_.Name)"
       }
     }
+    Write-Host ""
+    if ($Config.project.emulator.raine.config) {
+      Write-Host "raine configs for the project $($Config.project.name) :"
 
+      $Config.project.emulator.raine.config.ChildNodes | ForEach-Object {
+        Write-Host "run:raine:$($_.Name)"
+      }
+    }
+    Write-Host ""
+    Write-Host "--version"
+    Write-Host "--help"
     exit 1
 }